+++ provision/variables.tf
@@ -90,6 +90,17 @@ variable "instance_flex_ocpus" {
   default     = 1
 }
 
+variable "instance_state" {
+  type        = string
+  description = "(Updatable) The target state for the instance. Could be set to RUNNING or STOPPED."
+  default     = "RUNNING"
+
+  validation {
+    condition     = contains(["RUNNING", "STOPPED"], var.instance_state)
+    error_message = "Accepted values are RUNNING or STOPPED."
+  }
+}
+
 variable "shape" {
   description = "The shape of an instance."
   type        = string